remove utils.py. Sample models created in example files
